requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.1
boto3==1.26.10
botocore==1.29.10
Django==4.1.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-storages==1.13.1
django-summernote==0.8.20.0
djangorestframework==3.14.0
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.5
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.5
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.12
webencodings==0.5.1
whitenoise==6.2.0
